feat: include metadata and og to agenda
